requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.8
whitenoise
psycopg2-binary==2.9.10
geopy==2.4.0
gunicorn
# Optional (if needed for environment management)
python-dotenv==1.0.1
django-celery-beat
# For handling Excel files, if explicitly required
openpyxl==3.1.5

# For advanced CSV/Excel-like data processing
pandas==2.2.3
dj-database-url

# For JWT handling (if you are using token-based auth)
PyJWT==2.9.0
celery[redis]
redis
